* development:
  Return 404 when place not found
  Return 500 when not able to create favourite
  Return error when place not found
  new 404-response because not found
  Deleted getPlaceById-function. Gives problem because use of Response

+++ app/Http/Controllers/PlaceController.php
@@ -89,19 +89,27 @@ class PlaceController extends Controller
         try{
             $user = Auth::user();
             $place = Place::find($id);
+            if (!$place)
+            {
+                return Response('We weren\'t able to find the place', 404);
+            }
             $place->isFavouriteFrom()->attach($user);
             $place->save();
             return (new Response('Succesfully added the place to your favourites.',200));
         } catch(Exception $ex){
             Log::error($ex);
-            return 'We were not able to add this place to your favourites.';
+            return new Response('We were not able to add this place to your favourites.', 500);
         }
     }
 
     public function removeFromFavourites($id){
         try {
             $user = Auth::user();
-            $place = App\Place::find($id);
+            $place = Place::find($id);
+            if (!$place)
+            {
+                return new Response('We weren\'t able to find the place', 404);
+            }
             $place->isFavouriteFrom()->detach($user);
             $place->save();
             return (new Response('Successfully removed the place from your favourites.',200));
